refactor(cli): split read-view handlers by owner
