+++ docs/topology-status.md
@@ -28,10 +28,10 @@ Generated by `devtools render topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 702
+- **Stable** (no move scoped): 711
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 16
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 843
-- **Realized polylogue/**/*.py**: 843 files declared
+- **Total declared**: 852
+- **Realized polylogue/**/*.py**: 852 files declared
 <!-- END GENERATED: topology-status -->
