+++ polylogue/cli/read_views/__init__.py
@@ -1,0 +1,1 @@
+"""Executable CLI read-view handlers."""
